Pass eager cursors into an ODM\Cursor object
This deprecates the EagerCursor class since its main goal is already handled by
Doctrine\MongoDB\EagerCursor. This also allows us to fix a long-standing issue
where calling getSingleResult() on a query with reference priming enabled would
prime references for the entire dataset.
