+++ lib/Doctrine/ODM/MongoDB/EagerCursor.php
@@ -19,230 +19,14 @@
 
 namespace Doctrine\ODM\MongoDB;
 
-use Doctrine\MongoDB\EagerCursor as BaseEagerCursor;
-use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
-use Doctrine\ODM\MongoDB\Query\Query;
-use Doctrine\ODM\MongoDB\Query\ReferencePrimer;
-
 /**
  * EagerCursor wraps a Cursor instance and fetches all of its results upon
  * initialization.
  *
  * @since  1.0
  * @author Jonathan H. Wage <jonwage@gmail.com>
+ * @deprecated Deprecated in favor of using Cursor; will be removed in 2.0
  */
-class EagerCursor extends BaseEagerCursor
+class EagerCursor extends Cursor
 {
-    /**
-     * The ClassMetadata instance for the document class being queried.
-     *
-     * @var ClassMetadata
-     */
-    private $class;
-
-    /**
-     * Whether to hydrate results as document class instances.
-     *
-     * @var boolean
-     */
-    private $hydrate = true;
-
-    /**
-     * The UnitOfWork instance used for result hydration and preparing arguments
-     * for {@link Cursor::sort()}.
-     *
-     * @var UnitOfWork
-     */
-    private $unitOfWork;
-
-    /**
-     * Hints for UnitOfWork behavior.
-     *
-     * @var array
-     */
-    private $unitOfWorkHints = array();
-
-    /**
-     * ReferencePrimer object for priming references
-     *
-     * @var ReferencePrimer
-     */
-    private $referencePrimer;
-
-    /**
-     * Primers
-     *
-     * @var array
-     */
-    private $primers = array();
-
-    /**
-     * Whether references have been primed
-     *
-     * @var bool
-     */
-    private $referencesPrimed = false;
-
-    /**
-     * Constructor.
-     *
-     * @param Cursor        $cursor      Cursor instance being wrapped
-     * @param UnitOfWork    $unitOfWork  UnitOfWork for result hydration and query preparation
-     * @param ClassMetadata $class       ClassMetadata for the document class being queried
-     */
-    public function __construct(Cursor $cursor, UnitOfWork $uow, ClassMetadata $class)
-    {
-        $this->cursor = $cursor;
-        $this->unitOfWork = $uow;
-        $this->class = $class;
-    }
-
-    /**
-     * @see \Doctrine\MongoDB\EagerCursor::current()
-     * @see http://php.net/manual/en/iterator.current.php
-     */
-    public function current()
-    {
-        $this->primeReferences();
-
-        $current = parent::current();
-
-        if ($current === null || ! $this->hydrate) {
-            return $current;
-        }
-
-        return $this->unitOfWork->getOrCreateDocument($this->class->name, $current, $this->unitOfWorkHints);
-    }
-
-    /**
-     * Get hints for UnitOfWork behavior.
-     *
-     * @return array
-     */
-    public function getHints()
-    {
-        return $this->unitOfWorkHints;
-    }
-
-    /**
-     * Set hints for UnitOfWork behavior.
-     *
-     * @param array $hints
-     */
-    public function setHints(array $hints)
-    {
-        $this->unitOfWorkHints = $hints;
-    }
-
-    /**
-     * Set whether to hydrate results as document class instances.
-     *
-     * @see Cursor::hydrate()
-     * @param boolean $hydrate
-     * @return self
-     */
-    public function hydrate($hydrate = true)
-    {
-        $this->hydrate = (boolean) $hydrate;
-        return $this;
-    }
-
-    /**
-     * Initialize the internal data by converting the Cursor to an array.
-     */
-    public function initialize()
-    {
-        if ($this->initialized === false) {
-            $this->data = $this->cursor->getBaseCursor()->toArray(false);
-        }
-        $this->initialized = true;
-    }
-
-    /**
-     * Set whether to refresh hydrated documents that are already in the
-     * identity map.
-     *
-     * This option has no effect if hydration is disabled.
-     *
-     * @see Cursor::refresh()
-     * @param boolean $refresh
-     * @return self
-     */
-    public function refresh($refresh = true)
-    {
-        $this->unitOfWorkHints[Query::HINT_REFRESH] = (boolean) $refresh;
-        return $this;
-    }
-
-    /**
-     * Wrapper method for MongoCursor::setReadPreference().
-     *
-     * @see Cursor::setReadPreference()
-     * @param string $readPreference
-     * @param array  $tags
-     * @return self
-     */
-    public function setReadPreference($readPreference, array $tags = null)
-    {
-        $this->cursor->setReadPreference($readPreference, $tags);
-        $this->unitOfWorkHints[Query::HINT_READ_PREFERENCE] = $readPreference;
-        $this->unitOfWorkHints[Query::HINT_READ_PREFERENCE_TAGS] = $tags;
-        return $this;
-    }
-
-    /**
-     * Wrapper method for MongoCursor::slaveOkay().
-     *
-     * @see Cursor::slaveOkay()
-     * @param boolean $ok
-     * @return self
-     */
-    public function slaveOkay($ok = true)
-    {
-        $ok = (boolean) $ok;
-        $this->cursor->slaveOkay($ok);
-        $this->unitOfWorkHints[Query::HINT_SLAVE_OKAY] = $ok;
-        return $this;
-    }
-
-    /**
-     * @see \Doctrine\MongoDB\EagerCursor::toArray()
-     * @see \Doctrine\MongoDB\Iterator::toArray()
-     */
-    public function toArray()
-    {
-        $this->initialize();
-        return iterator_to_array($this);
-    }
-
-    /**
-     * @param array $primers
-     * @param ReferencePrimer $referencePrimer
-     * @return self
-     */
-    public function enableReferencePriming(array $primers, ReferencePrimer $referencePrimer)
-    {
-        $this->referencePrimer = $referencePrimer;
-        $this->primers = $primers;
-        return $this;
-    }
-
-    /**
-     * Prime references
-     */
-    protected function primeReferences()
-    {
-        if ($this->referencesPrimed || !$this->hydrate || empty($this->primers)) {
-            return;
-        }
-
-        $this->referencesPrimed = true;
-
-        foreach ($this->primers as $fieldName => $primer) {
-            $primer = is_callable($primer) ? $primer : null;
-            $this->referencePrimer->primeReferences($this->class, $this, $fieldName, $this->unitOfWorkHints, $primer);
-        }
-
-        $this->rewind();
-    }
 }
